feat: add check for each user's handler

+++ authentication-service/src/controllers/v0/userController.ts
@@ -4,11 +4,16 @@ import dotenv from 'dotenv';
 import jwt from "jsonwebtoken";
 import { login, register, deleteInputUser } from "./utils/auth";
 import { tokenExpiration, verifyToken } from "./utils/jwt";
-import { API_KEY_FIELD, USER_EMAIL_FIELD, USER_PASSWORD_FIELD, USER_JWT_TOKEN_EXPIRATION_FIELD, USER_JWT_TOKEN_FIELD, ADMIN_USER, NORMAL_USER, ERROR_HEADER } from "../../models/v0/headers/userHeaders";
+import { API_KEY_FIELD, USER_EMAIL_FIELD, USER_PASSWORD_FIELD, USER_JWT_TOKEN_EXPIRATION_FIELD, USER_JWT_TOKEN_FIELD, ADMIN_USER, NORMAL_USER, ERROR_HEADER, USER_ACTION_FIELD } from "../../models/v0/headers/userHeaders";
+import { AUTHENTICATE, DELETE, LOGIN, REGISTER } from "./utils/userActions";
 
 dotenv.config();
 
 const secretKey = process.env.SECRET_API_KEY || "__"
+
+function checkAction(tag: string, body: any, equalTo: string): boolean {
+    return body[tag] === equalTo;
+}
 
 function isAdmin(data:any): boolean { 
     if (API_KEY_FIELD in data) {
@@ -24,7 +29,7 @@ function fromBody<X>(body: any, key: string, defaultValue: X): X {
 
 const loginUser = async (request: Request, response: Response) => {
     const modelData = request.body;
-    if (modelData) {
+    if (modelData && checkAction(USER_ACTION_FIELD, modelData, LOGIN)) {
         response = await login(fromBody<string>(modelData, USER_EMAIL_FIELD, ""), fromBody<string>(modelData, USER_PASSWORD_FIELD, ""), response);
         response.end();
     }
@@ -32,7 +37,7 @@ const loginUser = async (request: Request, response: Response) => {
 
 const loginAdmin = async (request: Request, response: Response) => {
     const modelData = request.body;
-    if(modelData) {
+    if(modelData && checkAction(USER_ACTION_FIELD, modelData, LOGIN)) {
         if (isAdmin(modelData)) {
             response = await login(fromBody<string>(modelData, USER_EMAIL_FIELD, ""), fromBody<string>(modelData, USER_PASSWORD_FIELD, ""), response);
         }else {
@@ -44,7 +49,7 @@ const loginAdmin = async (request: Request, response: Response) => {
 
 const registerUser = async (request: Request, response: Response) => {
     const modelData = request.body;
-    if (modelData) {
+    if (modelData && checkAction(USER_ACTION_FIELD, modelData, REGISTER)) {
         response = await register(fromBody<string>(modelData, USER_EMAIL_FIELD, ""), fromBody<string>(modelData, USER_PASSWORD_FIELD, ""), NORMAL_USER, response);
         response.end()
     }
@@ -52,7 +57,8 @@ const registerUser = async (request: Request, response: Response) => {
 
 const registerAdmin = async (request: Request, response: Response) => {
     const modelData = request.body;
-    if (modelData) {
+
+    if (modelData && checkAction(USER_ACTION_FIELD, modelData, REGISTER)) {
         if (isAdmin(modelData)) {
             response = await register(fromBody<string>(modelData, USER_EMAIL_FIELD, ""), fromBody<string>(modelData, USER_PASSWORD_FIELD, ""), ADMIN_USER, response);
         }else {
@@ -64,7 +70,7 @@ const registerAdmin = async (request: Request, response: Response) => {
 
 const deleteUser = async (request: Request, response: Response) => {
     const modelData = request.body;
-    if (modelData) {
+    if (modelData && checkAction(USER_ACTION_FIELD, modelData, DELETE)) {
         response = await deleteInputUser(fromBody<string>(modelData, USER_EMAIL_FIELD, ""), response);
     }
     response.end();
@@ -85,7 +91,7 @@ const deleteAdmin = async (request: Request, response: Response) => {
 const checkToken = async (request: Request, response: Response) => {
     const modelData = request.body;
     try {
-        if (modelData) {
+        if (modelData && checkAction(USER_ACTION_FIELD, modelData, AUTHENTICATE)) {
             const jwtToken: string = fromBody<string>(modelData, USER_JWT_TOKEN_FIELD, "");
             const verified = await verifyToken(jwtToken);
             if (verified) {
